refactor: Teacher Filters Card
